feat: add style for post
