+++ vite-project/src/components/Form/Textarea.scss
@@ -24,5 +24,6 @@
 }
 
 .highlight {
-  color: navy;
+  color: cadetblue;
+  font-weight: bold;
 }
