+++ vite-project/src/components/Post/Post.module.scss
@@ -1,0 +1,32 @@
+.post {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  padding: 10px;
+  border-radius: 10px;
+  border: 1px solid rgb(90, 165, 250);
+  width: 300px;
+}
+
+.controls {
+  width: 100%;
+  display: flex;
+  justify-content: space-around;
+}
+
+.button {
+  width: 30%;
+  padding: 5px;
+  border: none;
+  border-radius: 10px;
+  color: white;
+  background-color: rgb(90, 165, 250);
+  margin: 0 auto;
+  cursor: pointer;
+  opacity: .8;
+  transition: all .2s;
+
+  &:hover {
+    opacity: 1;
+  }
+}
