+++ vite-project/src/components/UI/modal/MyModal.module.scss
@@ -6,6 +6,7 @@
   display: none;
   position: fixed;
   background-color: rgba(0, 0, 0, .5);
+  z-index: 2;
 }
 
 .myModal.active {
